#144 Move Elements data model creation in tests to separate helper class
