+++ Kentico.Kontent.Management.Tests/Unit/Data/ElementsData/ElementsData.json
@@ -1,0 +1,149 @@
+{
+  "elements": [
+    {
+      "element": {
+        "id": "ba7c8840-bcbc-5e3b-b292-24d0a60f3977"
+      },
+      "value": "On Roasts"
+    },
+    {
+      "searchable_value": "Almighty form!",
+      "element": {
+        "id": "47bf7d6d-285d-4ed0-9919-1d1a98b43acd"
+      },
+      "value": "{\"formId\": 42}"
+    },
+    {
+      "element": {
+        "id": "773940f4-9e67-4a26-a93f-67e55fd7d837"
+      },
+      "value": 3.14
+    },
+    {
+      "element": {
+        "id": "53a25074-b136-4a1f-a16d-3c130f696c66"
+      },
+      "value": [
+        {
+          "id": "00c0f86a-7c51-4e60-abeb-a150e9092e53"
+        },
+        {
+          "id": "8972dc90-ae2e-416e-995d-95df6c77e3b2"
+        }
+      ]
+    },
+    {
+      "element": {
+        "id": "9c6a4fbc-3f73-585f-9521-8d57636adf56"
+      },
+      "value": [
+        {
+          "id": "5c08a538-5b58-44eb-81ef-43fb37eeb815"
+        }
+      ]
+    },
+    {
+      "element": {
+        "id": "0827e079-3754-5a1d-9381-8ff695a5bbf7"
+      },
+      "value": "2017-07-04T00:00:00Z"
+    },
+    {
+      "element": {
+        "id": "15517aa3-da8a-5551-a4d4-555461fd5226"
+      },
+      "value": "Summary"
+    },
+    {
+      "components": [
+        {
+          "id": "46c05bd9-d418-4507-836c-9accc5a39db3",
+          "type": {
+            "id": "17ff8a28-ebe6-5c9d-95ea-18fe1ff86d2d"
+          },
+          "elements": [
+            {
+              "element": {
+                "id": "f09fb430-2a58-59cf-be03-621e1c367501"
+              },
+              "value": "https://twitter.com/ChrastinaOndrej/status/1417105245935706123"
+            },
+            {
+              "element": {
+                "id": "05017deb-18b2-5094-b367-e7f8796dd1b8"
+              },
+              "value": [
+                {
+                  "id": "061e69f7-0965-5e37-97bc-29963cfaebe8"
+                }
+              ]
+            },
+            {
+              "element": {
+                "id": "19e38194-a3a8-5d17-abed-9b70d7a5fd25"
+              },
+              "value": [
+                {
+                  "id": "dd78b09e-4337-599c-9701-20a0a165c63b"
+                }
+              ]
+            }
+          ]
+        }
+      ],
+      "element": {
+        "id": "55a88ab3-4009-5bf9-a590-f32162f09b92"
+      },
+      "value": "<h1>Light Roasts</h1><p>Usually roasted for 6 - 8 minutes or simply until achieving a light brown color.This method is used for milder coffee varieties and for coffee tasting.This type of roasting allows the natural characteristics of each coffee to show.The aroma of coffees produced from light roasts is usually more intense.The cup itself is more acidic and the concentration of caffeine is higher.</p><object type=\"application/kenticocloud\" data-type=\"component\" data-id=\"46c05bd9-d418-4507-836c-9accc5a39db3\"></object>"
+    },
+    {
+      "element": {
+        "id": "77108990-3c30-5ffb-8dcd-8eb85fc52cb1"
+      },
+      "value": [
+        {
+          "id": "4b628214-e4fe-4fe0-b1ff-955df33e1515"
+        }
+      ]
+    },
+    {
+      "element": {
+        "id": "0ee20a72-0aaa-521f-8801-df3d9293b7dd"
+      },
+      "value": "MetaKeywords"
+    },
+    {
+      "element": {
+        "id": "c1dc36b5-558d-55a2-8f31-787430a68e4d"
+      },
+      "value": [
+        {
+          "id": "6e8b18d5-c5e3-5fc1-9014-44c18ef5f5d8"
+        }
+      ]
+    },
+    {
+      "element": {
+        "id": "7df0048f-eaaf-50f8-85cf-fa0fc0d6d815"
+      },
+      "value": "MetaDescription"
+    },
+    {
+      "mode": "custom",
+      "element": {
+        "id": "1f37e15b-27a0-5f48-b314-03b401c19cee"
+      },
+      "value": "on-roasts"
+    },
+    {
+      "element": {
+        "id": "a29858ff-fa9f-5841-a682-d7fb6cc6effe"
+      },
+      "value": [
+        {
+          "id": "4b628214-e4fe-4fe0-b1ff-955df33e1515"
+        }
+      ]
+    }
+  ]
+}
